+++ EssentailFeed/EssentailFeed/Feed Cache/Infrastructure/CoreData/ManagedFeedImage.swift
@@ -1,0 +1,32 @@
+//
+//  ManagedFeedImage.swift
+//  EssentailFeed
+//
+//  Created by Vishal Wagh on 08/07/22.
+//
+
+import CoreData
+
+@objc(ManagedFeedImage)
+internal class ManagedFeedImage: NSManagedObject {
+    @NSManaged var id: UUID
+    @NSManaged var imageDescription: String?
+    @NSManaged var location: String?
+    @NSManaged var url: URL
+    @NSManaged var cache: ManagedCache
+    
+    internal static func images(from localFeed: [LocalFeedImage], in context: NSManagedObjectContext) -> NSOrderedSet{
+        return NSOrderedSet(array: localFeed.map({ local in
+            let managed = ManagedFeedImage(context: context)
+            managed.id = local.id
+            managed.imageDescription = local.description
+            managed.location = local.location
+            managed.url = local.url
+            return managed
+        }))
+    }
+    
+    internal var local: LocalFeedImage {
+        return LocalFeedImage(id: id, description: imageDescription, location: location, url: url)
+    }
+}
